更新

+++ TowerDefenseTutorial/Assets/Scripts/Enemy.cs
@@ -5,7 +5,7 @@ public class Enemy : MonoBehaviour
 
     public float speed = 10f;
 
-    public int health = 100;
+    public float health = 100f;
 
     public int value = 50;
 
@@ -19,7 +19,7 @@ public class Enemy : MonoBehaviour
         target = Waypoints.points[0];
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         health -= amount;
